Unify measure bundle helper tests to one file
